refactor : 폴더 위치 변경 (구조 변경 및 전체적인 리팩터링을 위해 따로 분리시킴)
